+++ src/main/java/stelnet/commodity/data/TableCellHelper.java
@@ -1,4 +1,4 @@
-package stelnet.commodity.extractor;
+package stelnet.commodity.data;
 
 import java.awt.Color;
