+++ src/main/java/stelnet/commodity/view/CommodityViewFactory.java
@@ -1,4 +1,4 @@
-package stelnet.commodity.element;
+package stelnet.commodity.view;
 
 import java.util.Arrays;
 import java.util.List;
@@ -8,11 +8,11 @@ import com.fs.starfarer.api.campaign.econ.MarketAPI;
 import org.lwjgl.input.Keyboard;
 
 import stelnet.commodity.CommodityBoard.CommodityTab;
-import stelnet.commodity.extractor.BuyMarketFactory;
-import stelnet.commodity.extractor.BuyTableContent;
-import stelnet.commodity.extractor.MarketFactory;
-import stelnet.commodity.extractor.SellMarketFactory;
-import stelnet.commodity.extractor.SellTableContent;
+import stelnet.commodity.data.BuyMarketFactory;
+import stelnet.commodity.data.BuyTableContent;
+import stelnet.commodity.data.MarketFactory;
+import stelnet.commodity.data.SellMarketFactory;
+import stelnet.commodity.data.SellTableContent;
 import stelnet.ui.Renderable;
 import stelnet.ui.Row;
 import stelnet.ui.Stack;
