+++ src/main/java/stelnet/commodity/view/IntelButton.java
@@ -1,4 +1,4 @@
-package stelnet.commodity.element;
+package stelnet.commodity.view;
 
 import java.awt.event.KeyEvent;
