+++ src/main/java/stelnet/filter/market/CommodityAvailableFilter.java
@@ -3,7 +3,7 @@ package stelnet.filter.market;
 import com.fs.starfarer.api.campaign.econ.CommodityOnMarketAPI;
 import com.fs.starfarer.api.campaign.econ.MarketAPI;
 
-import stelnet.commodity.extractor.TableCellHelper;
+import stelnet.commodity.data.TableCellHelper;
 
 public class CommodityAvailableFilter implements MarketFilter {
