+++ src/main/java/stelnet/storage/StorageBoard.java
@@ -13,8 +13,8 @@ import com.fs.starfarer.api.util.Misc;
 import stelnet.helper.GlobalHelper;
 import stelnet.helper.IntelHelper;
 import stelnet.helper.StorageHelper;
-import stelnet.storage.element.ControlViewFactory;
-import stelnet.storage.element.DisplayViewFactory;
+import stelnet.storage.view.ControlViewFactory;
+import stelnet.storage.view.DisplayViewFactory;
 import stelnet.ui.Callable;
 import stelnet.ui.GridRenderer;
 import stelnet.ui.Size;
